FHIR-parquet-explore/requirements.txt
azure-identity
pyyaml
pyarrow
pandas
numpy
pandas